Revert "Artist matching should now be a little bit more accurate."
