+++ src/backend/genius/artist.js
@@ -1,22 +1,17 @@
+const secrets = require('../secrets.json');
 const utils = require('../utils');
 const appleMusicApi = require('../appleMusicApi');
-const { stringComparator, geniusAxios } = require('../utils');
+
 const axios = require('axios');
 
-async function findMatch(hits, name) {
-  if (hits.length === 0) {
+async function findMatch(hits) {
+  if(hits.length === 0) {
     return null;
   }
 
   for (const hit of hits) {
-    if (hit.type === 'artist' && stringComparator(hit.result.name, name)) {
+    if (hit.type === "artist") {
       return hit.result;
-    }
-  }
-
-  for (const hit of hits) {
-    if (hit.type === 'artist') {
-      return hit.result; // Returns the first artist it finds, if it cannot find one which directly matches the name.
     }
   }
 
@@ -24,7 +19,11 @@ async function findMatch(hits, name) {
 }
 
 async function fetchArtist(hit) {
-  const { data } = await geniusAxios.get(hit.api_path);
+  const {data} = await axios.get(`https://api.genius.com${hit.api_path}`, {
+    headers: {
+      Authorization: `Bearer ${secrets.GENIUS_API_KEY}`,
+    }
+  });
 
   const artist = data.response.artist;
   delete artist.current_user_metadata;
@@ -56,7 +55,7 @@ async function handle({ artistId }) {
     return null;
   }
 
-  const match = await findMatch(hits, artistName);
+  const match = await findMatch(hits);
 
   if (match) {
     return await fetchArtist(match);
@@ -66,7 +65,7 @@ async function handle({ artistId }) {
 }
 
 module.exports = {
-  async details(event) {
+  details: async function(event) {
     const params = event.queryStringParameters;
 
     try {
@@ -76,5 +75,5 @@ module.exports = {
     } catch (e) {
       return utils.generateError(500, e);
     }
-  },
+  }
 };
